Refactor: PointHistory내 필터 기능 filter.tsx로 분리
